get_game_stats

This function fetches detailed statistics for a game given the game ID. It fetches ten variables that are accessed as lists with home (h) and away (a) players: goals_scored, assists, own_goals, penalties_saved, yellow_cards, red_cards, saves, bonus, and bps.

get_game_stats(gameid = NULL)

Arguments

gameid
The game ID. Can be found with get_game_list().

Examples

get_game_stats(1)

#>          identifier
#> 1      goals_scored
#> 2           assists
#> 3         own_goals
#> 4   penalties_saved
#> 5  penalties_missed
#> 6      yellow_cards
#> 7         red_cards
#> 8             saves
#> 9             bonus
#> 10              bps
#>                                                                                                                         a
#> 1                                                                                                                    NULL
#> 2                                                                                                                    NULL
#> 3                                                                                                                    NULL
#> 4                                                                                                                    NULL
#> 5                                                                                                                    NULL
#> 6                                                                                                                  1, 375
#> 7                                                                                                                    NULL
#> 8                                                                                                                  2, 363
#> 9                                                                                                                    NULL
#> 10 19, 17, 15, 15, 13, 12, 12, 11, 5, 4, 4, 4, 3, 1, 369, 363, 362, 370, 364, 367, 375, 376, 365, 361, 379, 382, 377, 366
#>                                                                                                                 h
#> 1                                                                                                          1, 141
#> 2                                                                                                          1, 137
#> 3                                                                                                            NULL
#> 4                                                                                                            NULL
#> 5                                                                                                            NULL
#> 6                                                                                                  1, 1, 131, 142
#> 7                                                                                                            NULL
#> 8                                                                                                          5, 128
#> 9                                                                                          3, 2, 1, 128, 137, 486
#> 10 31, 26, 24, 23, 22, 18, 14, 6, 5, 3, 3, 3, -7, 128, 137, 486, 146, 134, 141, 130, 138, 131, 133, 142, 489, 140