get_player_details

This function fetches detailed, gameweek-by-gameweek information for the current season on selected players. If parameters left blank, the function fetches all current players - note that takes a while to load.

get_player_details(playerid = NULL, name = NULL, season = NULL)

Arguments

playerid
The player's ID. Can be found using get_player_id(). You can list multiple, e.g. with c(). Won't work with prior season's data.

name
Alternatively, the player's full name, as listed on the official Fantasy Premier League site (for instance: "Richarlison de Andrade", not "Richarlison"). If blank, the function fetches all players.

season
To retrieve prior season data, enter a 2-digit year corresponding to the start of the historical FPL season of interest (e.g. '17' for the 2017/2018 season).

Examples

get_player_details(300)

#>    element fixture opponent_team total_points was_home         kickoff_time
#> 1      300       2             3            0    FALSE 2019-08-10T14:00:00Z
#> 2      300      18             7            0     TRUE 2019-08-18T13:00:00Z
#> 3      300      27             9            0     TRUE 2019-08-24T14:00:00Z
#> 4      300      33             6            0    FALSE 2019-08-31T14:00:00Z
#> 5      300      47            16            0     TRUE 2019-09-14T14:00:00Z
#> 6      300      55             8            0    FALSE 2019-09-21T14:00:00Z
#> 7      300      68            10            0     TRUE 2019-09-28T11:30:00Z
#> 8      300      79            18            0    FALSE 2019-10-05T14:00:00Z
#> 9      300      88             1            0     TRUE 2019-10-21T19:00:00Z
#> 10     300     100            19            0    FALSE 2019-10-26T14:00:00Z
#> 11     300     108             5            0     TRUE 2019-11-02T15:00:00Z
#> 12     300     119            17            0    FALSE 2019-11-09T15:00:00Z
#> 13     300     128            12            0     TRUE 2019-11-24T16:30:00Z
#> 14     300     140            20            0    FALSE 2019-12-01T14:00:00Z
#> 15     300     145            13            0     TRUE 2019-12-05T19:30:00Z
#> 16     300     157            14            0    FALSE 2019-12-08T14:00:00Z
#> 17     300     168             2            0     TRUE 2019-12-14T15:00:00Z
#> 18     300     173             4            0    FALSE 2019-12-21T15:00:00Z
#> 19     300     188            18            0     TRUE 2019-12-26T15:00:00Z
#> 20     300     195            11            0    FALSE 2019-12-29T18:00:00Z
#> 21     300     204            10            0    FALSE 2020-01-02T20:00:00Z
#> 22     300     218            19            0     TRUE 2020-01-10T20:00:00Z
#> 23     300     221             1            0    FALSE 2020-01-18T15:00:00Z
#> 24     300     236            11            0     TRUE 2020-01-21T19:30:00Z
#> 25     300     243             7            0    FALSE 2020-02-01T15:00:00Z
#> 26     300     258             3            0     TRUE 2020-02-09T14:00:00Z
#> 27     300     268             4            0     TRUE 2020-02-22T15:00:00Z
#> 28     300     288            14            0     TRUE 2020-03-07T15:00:00Z
#> 29     300     271             2            0    FALSE 2020-06-17T17:00:00Z
#> 30     300     296            13            0    FALSE 2020-06-21T13:00:00Z
#> 31     300     305            12            0    FALSE 2020-06-24T17:00:00Z
#>    team_h_score team_a_score round minutes goals_scored assists clean_sheets
#> 1             1            1     1       0            0       0            0
#> 2             1            0     2       0            0       0            0
#> 3             1            2     3       0            0       0            0
#> 4             2            2     4       0            0       0            0
#> 5             0            1     5       0            0       0            0
#> 6             0            2     6       0            0       0            0
#> 7             0            1     7       0            0       0            0
#> 8             0            0     8       0            0       0            0
#> 9             1            0     9       0            0       0            0
#> 10            1            1    10       0            0       0            0
#> 11            3            0    11       0            0       0            0
#> 12            1            1    12       0            0       0            0
#> 13            3            3    13       0            0       0            0
#> 14            1            1    14       0            0       0            0
#> 15            0            2    15       0            0       0            0
#> 16            1            2    16       0            0       0            0
#> 17            2            0    17       0            0       0            0
#> 18            0            1    18       0            0       0            0
#> 19            1            1    19       0            0       0            0
#> 20            2            0    20       0            0       0            0
#> 21            2            0    21       0            0       0            0
#> 22            1            0    22       0            0       0            0
#> 23            1            1    23       0            0       0            0
#> 24            0            1    24       0            0       0            0
#> 25            0            1    25       0            0       0            0
#> 26            2            1    26       0            0       0            0
#> 27            1            1    27       0            0       0            0
#> 28            1            0    29       0            0       0            0
#> 29            0            0    39       0            0       0            0
#> 30            3            0    39       0            0       0            0
#> 31            3            0    40       0            0       0            0
#>    goals_conceded own_goals penalties_saved penalties_missed yellow_cards
#> 1               0         0               0                0            0
#> 2               0         0               0                0            0
#> 3               0         0               0                0            0
#> 4               0         0               0                0            0
#> 5               0         0               0                0            0
#> 6               0         0               0                0            0
#> 7               0         0               0                0            0
#> 8               0         0               0                0            0
#> 9               0         0               0                0            0
#> 10              0         0               0                0            0
#> 11              0         0               0                0            0
#> 12              0         0               0                0            0
#> 13              0         0               0                0            0
#> 14              0         0               0                0            0
#> 15              0         0               0                0            0
#> 16              0         0               0                0            0
#> 17              0         0               0                0            0
#> 18              0         0               0                0            0
#> 19              0         0               0                0            0
#> 20              0         0               0                0            0
#> 21              0         0               0                0            0
#> 22              0         0               0                0            0
#> 23              0         0               0                0            0
#> 24              0         0               0                0            0
#> 25              0         0               0                0            0
#> 26              0         0               0                0            0
#> 27              0         0               0                0            0
#> 28              0         0               0                0            0
#> 29              0         0               0                0            0
#> 30              0         0               0                0            0
#> 31              0         0               0                0            0
#>    red_cards saves bonus bps influence creativity threat ict_index value
#> 1          0     0     0   0       0.0        0.0    0.0       0.0    55
#> 2          0     0     0   0       0.0        0.0    0.0       0.0    55
#> 3          0     0     0   0       0.0        0.0    0.0       0.0    54
#> 4          0     0     0   0       0.0        0.0    0.0       0.0    54
#> 5          0     0     0   0       0.0        0.0    0.0       0.0    53
#> 6          0     0     0   0       0.0        0.0    0.0       0.0    53
#> 7          0     0     0   0       0.0        0.0    0.0       0.0    53
#> 8          0     0     0   0       0.0        0.0    0.0       0.0    53
#> 9          0     0     0   0       0.0        0.0    0.0       0.0    53
#> 10         0     0     0   0       0.0        0.0    0.0       0.0    53
#> 11         0     0     0   0       0.0        0.0    0.0       0.0    53
#> 12         0     0     0   0       0.0        0.0    0.0       0.0    53
#> 13         0     0     0   0       0.0        0.0    0.0       0.0    53
#> 14         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 15         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 16         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 17         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 18         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 19         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 20         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 21         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 22         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 23         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 24         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 25         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 26         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 27         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 28         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 29         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 30         0     0     0   0       0.0        0.0    0.0       0.0    52
#> 31         0     0     0   0       0.0        0.0    0.0       0.0    52
#>    transfers_balance selected transfers_in transfers_out playername
#> 1                  0     6067            0             0 Mark Duffy
#> 2              -2198     3837           28          2226 Mark Duffy
#> 3               -998     2929           15          1013 Mark Duffy
#> 4               -664     2288           17           681 Mark Duffy
#> 5               -353     1955           20           373 Mark Duffy
#> 6               -161     1797            9           170 Mark Duffy
#> 7               -125     1674           20           145 Mark Duffy
#> 8                -84     1603            9            93 Mark Duffy
#> 9                -54     1548            9            63 Mark Duffy
#> 10               -33     1519            8            41 Mark Duffy
#> 11               -30     1492           12            42 Mark Duffy
#> 12               -22     1475           16            38 Mark Duffy
#> 13               -38     1443           17            55 Mark Duffy
#> 14               -23     1424           11            34 Mark Duffy
#> 15               -11     1417            3            14 Mark Duffy
#> 16               -15     1400            6            21 Mark Duffy
#> 17               -17     1390           11            28 Mark Duffy
#> 18                -8     1385           11            19 Mark Duffy
#> 19                -3     1388           17            20 Mark Duffy
#> 20                -9     1381            9            18 Mark Duffy
#> 21               -16     1367            8            24 Mark Duffy
#> 22               -13     1355            6            19 Mark Duffy
#> 23                -8     1349            8            16 Mark Duffy
#> 24               -11     1339            3            14 Mark Duffy
#> 25                -9     1335            8            17 Mark Duffy
#> 26                 1     1339            9             8 Mark Duffy
#> 27                -4     1334            7            11 Mark Duffy
#> 28                -3     1331           10            13 Mark Duffy
#> 29                -5     1327           42            47 Mark Duffy
#> 30                -5     1327           42            47 Mark Duffy
#> 31                 0     1331           10            10 Mark Duffy
#>  [ reached 'max' / getOption("max.print") -- omitted 7 rows ]

get_player_details(name="Virgil van Dijk")

#>    element fixture opponent_team total_points was_home         kickoff_time
#> 1      183       1            14            8     TRUE 2019-08-09T19:00:00Z
#> 2      183      19            16            2    FALSE 2019-08-17T14:00:00Z
#> 3      183      24             1            2     TRUE 2019-08-24T16:30:00Z
#> 4      183      32             5            7    FALSE 2019-08-31T16:30:00Z
#> 5      183      44            13            2     TRUE 2019-09-14T11:30:00Z
#> 6      183      53             6            2    FALSE 2019-09-22T15:30:00Z
#> 7      183      68            15            8    FALSE 2019-09-28T11:30:00Z
#> 8      183      74             9            2     TRUE 2019-10-05T14:00:00Z
#> 9      183      87            12            2    FALSE 2019-10-20T15:30:00Z
#> 10     183      94            17            2     TRUE 2019-10-27T16:30:00Z
#> 11     183     102             2            1    FALSE 2019-11-02T15:00:00Z
#> 12     183     114            11            2     TRUE 2019-11-10T16:30:00Z
#> 13     183     125             7            2    FALSE 2019-11-23T15:00:00Z
#> 14     183     134             4           17     TRUE 2019-11-30T15:00:00Z
#> 15     183     149             8            1     TRUE 2019-12-04T20:15:00Z
#> 16     183     152             3            6    FALSE 2019-12-07T15:00:00Z
#> 17     183     166            18            7     TRUE 2019-12-14T12:30:00Z
#> 18     183     186             9            6    FALSE 2019-12-26T20:00:00Z
#> 19     183     194            20            6     TRUE 2019-12-29T16:30:00Z
#> 20     183     204            15            7     TRUE 2020-01-02T20:00:00Z
#> 21     183     219            17            6    FALSE 2020-01-11T17:30:00Z
#> 22     183     224            12           15     TRUE 2020-01-19T16:30:00Z
#> 23     183     237            20            2    FALSE 2020-01-23T20:00:00Z
#> 24     183     180            19            6    FALSE 2020-01-29T19:45:00Z
#> 25     183     245            16            6     TRUE 2020-02-01T15:00:00Z
#> 26     183     257            14            6    FALSE 2020-02-15T17:30:00Z
#> 27     183     266            19            1     TRUE 2020-02-24T20:00:00Z
#> 28     183     279            18            1    FALSE 2020-02-29T17:30:00Z
#> 29     183     286             3            6     TRUE 2020-03-07T12:30:00Z
#> 30     183     294             8            6    FALSE 2020-06-21T18:00:00Z
#> 31     183     304             7            9     TRUE 2020-06-24T19:15:00Z
#>    team_h_score team_a_score round minutes goals_scored assists clean_sheets
#> 1             4            1     1      90            1       0            0
#> 2             1            2     2      90            0       0            0
#> 3             3            1     3      90            0       0            0
#> 4             0            3     4      90            0       0            1
#> 5             3            1     5      90            0       0            0
#> 6             1            2     6      90            0       0            0
#> 7             0            1     7      90            0       0            1
#> 8             2            1     8      90            0       0            0
#> 9             1            1     9      90            0       0            0
#> 10            2            1    10      90            0       0            0
#> 11            1            2    11      90            0       0            0
#> 12            3            1    12      90            0       0            0
#> 13            1            2    13      90            0       0            0
#> 14            2            1    14      90            2       0            0
#> 15            5            2    15      90            0       0            0
#> 16            0            3    16      90            0       0            1
#> 17            2            0    17      90            0       0            1
#> 18            0            4    19      90            0       0            1
#> 19            1            0    20      90            0       0            1
#> 20            2            0    21      90            0       0            1
#> 21            0            1    22      90            0       0            1
#> 22            2            0    23      90            1       0            1
#> 23            1            2    24      90            0       0            0
#> 24            0            2    24      90            0       0            1
#> 25            4            0    25      90            0       0            1
#> 26            0            1    26      90            0       0            1
#> 27            3            2    27      90            0       0            0
#> 28            3            0    28      90            0       0            0
#> 29            2            1    29      90            0       1            0
#> 30            0            0    39      90            0       0            1
#> 31            4            0    40      90            0       1            1
#>    goals_conceded own_goals penalties_saved penalties_missed yellow_cards
#> 1               1         0               0                0            0
#> 2               1         0               0                0            0
#> 3               1         0               0                0            0
#> 4               0         0               0                0            0
#> 5               1         0               0                0            0
#> 6               1         0               0                0            0
#> 7               0         0               0                0            0
#> 8               1         0               0                0            0
#> 9               1         0               0                0            0
#> 10              1         0               0                0            0
#> 11              1         0               0                0            1
#> 12              1         0               0                0            0
#> 13              1         0               0                0            0
#> 14              1         0               0                0            0
#> 15              2         0               0                0            0
#> 16              0         0               0                0            0
#> 17              0         0               0                0            0
#> 18              0         0               0                0            0
#> 19              0         0               0                0            0
#> 20              0         0               0                0            0
#> 21              0         0               0                0            0
#> 22              0         0               0                0            0
#> 23              1         0               0                0            0
#> 24              0         0               0                0            0
#> 25              0         0               0                0            0
#> 26              0         0               0                0            0
#> 27              2         0               0                0            0
#> 28              3         0               0                0            0
#> 29              1         0               0                0            0
#> 30              0         0               0                0            0
#> 31              0         0               0                0            0
#>    red_cards saves bonus bps influence creativity threat ict_index value
#> 1          0     0     0  27      52.4        0.8   21.0       7.4    65
#> 2          0     0     0  16      30.8        1.1    2.0       3.4    65
#> 3          0     0     0   7      11.0        0.8   20.0       3.2    65
#> 4          0     0     1  27      20.8        1.0    2.0       2.4    65
#> 5          0     0     0  17      24.4        4.5   29.0       5.8    65
#> 6          0     0     0  13      14.2        0.5    0.0       1.5    65
#> 7          0     0     2  29      31.0       14.1    2.0       4.7    65
#> 8          0     0     0  18      17.0        3.4    4.0       2.4    65
#> 9          0     0     0  13      16.4        1.3    0.0       1.8    65
#> 10         0     0     0  14      16.4        2.6   17.0       3.6    65
#> 11         0     0     0  10      18.0       11.8    4.0       3.4    65
#> 12         0     0     0  15      28.4        0.1    2.0       3.1    64
#> 13         0     0     0  15      23.0        4.7   56.0       8.4    64
#> 14         0     0     3  42      80.6        3.7   36.0      12.0    64
#> 15         0     0     0  15      18.0        0.3    0.0       1.8    64
#> 16         0     0     0  27      22.6        2.1    0.0       2.5    64
#> 17         0     0     1  27      16.2        5.0    0.0       2.1    64
#> 18         0     0     0  22      12.6        1.3    0.0       1.4    64
#> 19         0     0     0  29      18.8        1.3    0.0       2.0    64
#> 20         0     0     1  31      30.0        1.4    0.0       3.1    64
#> 21         0     0     0  22      23.4        0.9   33.0       5.7    64
#> 22         0     0     3  44      55.0        0.4   21.0       7.6    64
#> 23         0     0     0  18      19.8        0.2    2.0       2.2    64
#> 24         0     0     0  27      20.6        1.6    0.0       2.2    64
#> 25         0     0     0  23      22.0        0.4   17.0       3.9    65
#> 26         0     0     0  24       9.2        2.0   21.0       3.2    65
#> 27         0     0     0  14      24.6        2.8   44.0       7.1    65
#> 28         0     0     0  16      24.0       14.0    1.0       3.9    66
#> 29         0     0     1  29      37.8       32.1   21.0       9.1    66
#> 30         0     0     0  23      13.8        0.7   17.0       3.2    65
#> 31         0     0     0  30      14.6        6.8   22.0       4.3    65
#>    transfers_balance selected transfers_in transfers_out      playername
#> 1                  0  2725192            0             0 Virgil van Dijk
#> 2              60548  3217035       107209         46661 Virgil van Dijk
#> 3             -71028  3261181        94480        165508 Virgil van Dijk
#> 4             -62084  3306073        87396        149480 Virgil van Dijk
#> 5               6215  3401581       128944        122729 Virgil van Dijk
#> 6            -120067  3308891        32733        152800 Virgil van Dijk
#> 7            -111045  3234963        39184        150229 Virgil van Dijk
#> 8             -28162  3233186        35562         63724 Virgil van Dijk
#> 9             -72180  3185653        27864        100044 Virgil van Dijk
#> 10            -40156  3161305        24219         64375 Virgil van Dijk
#> 11           -160108  3008759        20403        180511 Virgil van Dijk
#> 12           -149272  2864573        10214        159486 Virgil van Dijk
#> 13            -72586  2821719        35815        108401 Virgil van Dijk
#> 14            -73511  2756210        18321         91832 Virgil van Dijk
#> 15             48461  2820753        60061         11600 Virgil van Dijk
#> 16             15023  2844566        52494         37471 Virgil van Dijk
#> 17             34448  2898864        68807         34359 Virgil van Dijk
#> 18             17629  2852799        46904         29275 Virgil van Dijk
#> 19              3931  2866541        48470         44539 Virgil van Dijk
#> 20             31116  2909838        82103         50987 Virgil van Dijk
#> 21             -1951  2922685        56997         58948 Virgil van Dijk
#> 22             24370  2959715        70071         45701 Virgil van Dijk
#> 23            197071  3177803       222215         25144 Virgil van Dijk
#> 24            197071  3177803       222215         25144 Virgil van Dijk
#> 25             80533  3265907       121745         41212 Virgil van Dijk
#> 26             37865  3308961        73854         35989 Virgil van Dijk
#> 27             30036  3350597        59837         29801 Virgil van Dijk
#> 28            -16357  3323775        30707         47064 Virgil van Dijk
#> 29            -44047  3291071        26563         70610 Virgil van Dijk
#> 30            -76296  3249515       132396        208692 Virgil van Dijk
#> 31             31356  3309746        55153         23797 Virgil van Dijk
#>  [ reached 'max' / getOption("max.print") -- omitted 7 rows ]