get_player_hist

This function fetches historical summary information on selected players given player ID(s).

get_player_hist(playerid = NULL)

Arguments

playerid
The player's ID. Can be found using get_player_id(). You can list multiple , e.g. with c(). If blank, the function fetches all current players - note that takes a while to load.

Examples

get_player_hist(playerid=200)

#>    season_name element_code start_cost end_cost total_points minutes
#> 1      2006/07        17745         35       35            0       0
#> 2      2007/08        17745         40       40           41     630
#> 3      2008/09        17745         40       39            2      72
#> 4      2014/15        17745         45       45           83    2160
#> 5      2015/16        17745         45       50          147    3420
#> 6      2016/17        17745         50       49          109    2666
#> 7      2017/18        17745         50       51          123    2965
#> 8      2018/19        17745         50       50          120    3420
#> 9      2019/20        17745         50       55          156    3420
#> 10     2020/21        17745         55       54          128    3420
#>    goals_scored assists clean_sheets goals_conceded own_goals penalties_saved
#> 1             0       0            0              0         0               0
#> 2             0       0            0              5         0               1
#> 3             0       0            0              1         0               0
#> 4             0       0            6             37         0               0
#> 5             0       0           15             36         0               0
#> 6             0       0            8             47         0               0
#> 7             0       0            8             47         1               2
#> 8             0       0           10             48         0               0
#> 9             0       0           13             41         0               2
#> 10            0       0           11             50         1               0
#>    penalties_missed yellow_cards red_cards saves bonus bps influence creativity
#> 1                 0            0         0     0     0   0       0.0        0.0
#> 2                 0            0         0    24     2   0       0.0        0.0
#> 3                 0            0         0     2     0   0       0.0        0.0
#> 4                 0            0         0    80     6 366       0.0        0.0
#> 5                 0            2         0   100     5 724       0.0        0.0
#> 6                 0            1         0   113    10 585     757.4        0.0
#> 7                 0            3         0    93    17 592     687.2       11.1
#> 8                 0            3         0    90     6 618     756.8       10.0
#> 9                 0            2         0    97    10 707     863.4       30.0
#> 10                0            0         0    88     6 631     769.0        0.0
#>    threat ict_index        playername
#> 1     0.0       0.0 Kasper Schmeichel
#> 2     0.0       0.0 Kasper Schmeichel
#> 3     0.0       0.0 Kasper Schmeichel
#> 4     0.0       0.0 Kasper Schmeichel
#> 5     0.0       0.0 Kasper Schmeichel
#> 6     0.0      75.8 Kasper Schmeichel
#> 7     0.0      69.9 Kasper Schmeichel
#> 8     4.0      76.9 Kasper Schmeichel
#> 9     2.0      89.6 Kasper Schmeichel
#> 10    0.0      76.9 Kasper Schmeichel

get_player_hist(playerid=get_player_id("Sead Kolasinac")$id)

#>   season_name element_code start_cost end_cost total_points minutes
#> 1     2017/18       111457         60       58          108    2146
#> 2     2018/19       111457         50       49           81    1886
#> 3     2019/20       111457         55       52           55    1694
#> 4     2020/21       111457         50       48            2      90
#>   goals_scored assists clean_sheets goals_conceded own_goals penalties_saved
#> 1            2       4           10             30         0               0
#> 2            0       7            5             29         1               0
#> 3            0       2            4             26         0               0
#> 4            0       0            0              1         0               0
#>   penalties_missed yellow_cards red_cards saves bonus bps influence creativity
#> 1                0            3         0     0     6 510     443.2      334.9
#> 2                0            5         0     0    10 382     323.2      445.9
#> 3                0            4         0     0     1 331     269.6      182.5
#> 4                0            0         0     0     0   9       3.6        5.4
#>   threat ict_index     playername
#> 1  404.0     118.3 Sead Kolasinac
#> 2  242.0     101.0 Sead Kolasinac
#> 3   81.0      53.0 Sead Kolasinac
#> 4    4.0       1.3 Sead Kolasinac