get_player_info

This function fetches selected summary information on selected players given full player name(s).

get_player_info(name = NULL)

Arguments

name
The player's full name, as listed on the official Fantasy Premier League site (for instance: "Richarlison de Andrade", not "Richarlison"). If blank, the function fetches all players.

Examples

get_player_info(name=c("Jesse Lingard","Aleksandar Mitrovic"))

#>    id      photo web_name team_code status   code first_name second_name
#> 1 480 128389.jpg Mitrovic        54      a 128389 Aleksandar    Mitrovic
#> 2 300 109322.jpg  Lingard         1      a 109322      Jesse     Lingard
#>   squad_number news now_cost           news_added chance_of_playing_this_round
#> 1            9            64                 <NA>                           NA
#> 2           14            67 2019-02-12T22:31:20Z                           75
#>   chance_of_playing_next_round value_form value_season cost_change_start
#> 1                           NA        0.3         18.1                -1
#> 2                          100        0.0         10.4                -3
#>   cost_change_event cost_change_start_fall cost_change_event_fall in_dreamteam
#> 1                 0                      1                      0        FALSE
#> 2                 0                      3                      0        FALSE
#>   dreamteam_count selected_by_percent form transfers_out transfers_in
#> 1               3                11.9  2.0       2506231      2717661
#> 2               1                 2.8  0.3        785094       623936
#>   transfers_out_event transfers_in_event loans_in loans_out loaned_in
#> 1               14000               1220        0         0         0
#> 2                 890               3688        0         0         0
#>   loaned_out total_points event_points points_per_game ep_this ep_next special
#> 1          0          116            2             3.7     0.5     2.0   FALSE
#> 2          0           70            0             3.2     0.0     1.6   FALSE
#>   minutes goals_scored assists clean_sheets goals_conceded own_goals
#> 1    2740           10       3            2             69         0
#> 2    1278            4       3            3             18         0
#>   penalties_saved penalties_missed yellow_cards red_cards saves bonus bps
#> 1               0                0            6         0     0    12 321
#> 2               0                0            2         0     0     5 236
#>   influence creativity threat ict_index ea_index element_type team
#> 1     575.6      424.4 1276.0     227.9        0            4    9
#> 2     272.0      238.4  323.0      83.2        0            3   14
#>            playername
#> 1 Aleksandar Mitrovic
#> 2       Jesse Lingard