get_player_details

This function fetches detailed, gameweek-by-gameweek information for the current season on selected players given player ID(s).

get_player_details(playerid = NULL)

Arguments

playerid
The player's ID. Can be found using get_player_id(). You can list multiple , e.g. with c(). If blank, the function fetches all current players - note that takes a while to load.

Examples

get_player_details(300)

#>       id         kickoff_time kickoff_time_formatted team_h_score team_a_score
#> 1    300 2018-08-10T19:00:00Z           10 Aug 20:00            2            1
#> 2    825 2018-08-19T15:00:00Z           19 Aug 16:00            3            2
#> 3   1352 2018-08-27T19:00:00Z           27 Aug 20:00            0            3
#> 4   1883 2018-09-02T15:00:00Z           02 Sep 16:00            0            2
#> 5   2420 2018-09-15T16:30:00Z           15 Sep 17:30            1            2
#> 6   2957 2018-09-22T14:00:00Z           22 Sep 15:00            1            1
#> 7   3495 2018-09-29T11:30:00Z           29 Sep 12:30            3            1
#> 8   4034 2018-10-06T16:30:00Z           06 Oct 17:30            3            2
#> 9   4576 2018-10-20T11:30:00Z           20 Oct 12:30            2            2
#> 10  5122 2018-10-28T16:00:00Z           28 Oct 16:00            2            1
#> 11  5670 2018-11-03T12:30:00Z           03 Nov 12:30            1            2
#> 12  6220 2018-11-11T16:30:00Z           11 Nov 16:30            3            1
#> 13  6772 2018-11-24T15:00:00Z           24 Nov 15:00            0            0
#> 14  7325 2018-12-01T17:30:00Z           01 Dec 17:30            2            2
#> 15  7881 2018-12-05T20:00:00Z           05 Dec 20:00            2            2
#> 16  8439 2018-12-08T15:00:00Z           08 Dec 15:00            4            1
#> 17  8998 2018-12-16T16:00:00Z           16 Dec 16:00            3            1
#> 18  9560 2018-12-22T17:30:00Z           22 Dec 17:30            1            5
#> 19 10125 2018-12-26T15:00:00Z           26 Dec 15:00            3            1
#> 20 10693 2018-12-30T16:30:00Z           30 Dec 16:30            4            1
#> 21 11261 2019-01-02T20:00:00Z           02 Jan 20:00            0            2
#> 22 11831 2019-01-13T16:30:00Z           13 Jan 16:30            0            1
#> 23 12405 2019-01-19T15:00:00Z           19 Jan 15:00            2            1
#> 24 12984 2019-01-29T20:00:00Z           29 Jan 20:00            2            2
#> 25 13613 2019-02-03T14:05:00Z           03 Feb 14:05            0            1
#> 26 14221 2019-02-09T12:30:00Z           09 Feb 12:30            0            3
#> 27 14730 2019-02-24T14:05:00Z           24 Feb 14:05            0            0
#> 28 15305 2019-02-27T20:00:00Z           27 Feb 20:00            1            3
#> 29 15910 2019-03-02T15:00:00Z           02 Mar 15:00            3            2
#> 30 16516 2019-03-10T16:30:00Z           10 Mar 16:30            2            0
#>    was_home round total_points value transfers_balance selected transfers_in
#> 1      TRUE     1            0    70                 0   283400            0
#> 2     FALSE     2            1    69            -75370   208433         1669
#> 3      TRUE     3            2    69            -33173   182285         3612
#> 4     FALSE     4            3    68             -9541   177382         8928
#> 5     FALSE     5            2    68              6415   188722        23617
#> 6      TRUE     6            2    68             -4176   186148        11036
#> 7     FALSE     7            0    68            -13769   173407         3961
#> 8      TRUE     8            0    68            -26916   146641          478
#> 9     FALSE     9            0    67            -24679   122439          129
#> 10     TRUE    10            1    67             -7056   115987          110
#> 11    FALSE    11            1    67             -2069   114352          642
#> 12    FALSE    12            1    67             -1421   113308          706
#> 13     TRUE    13            1    67              3798   118356         5866
#> 14    FALSE    14            1    67              -574   117624         1467
#> 15     TRUE    15            6    67             -2034   115795          682
#> 16     TRUE    16            2    67             10580   129550        12149
#> 17    FALSE    17            9    67             -3705   123898         2574
#> 18    FALSE    18           18    67             12599   138256        15321
#> 19     TRUE    19            5    68            282338   435318       288728
#> 20     TRUE    20            2    69             52528   483067        87111
#> 21    FALSE    21            1    69            -52595   430396        36850
#> 22    FALSE    22            3    68            -90242   344256        26524
#> 23     TRUE    23            2    68            -16879   331805        39764
#> 24     TRUE    24            4    68            -18800   314574        17937
#> 25    FALSE    25            2    67            -17682   294756        12742
#> 26    FALSE    26            0    67            -13252   285907        14315
#> 27     TRUE    27            1    67            -57778   221926         2501
#> 28    FALSE    28            0    67            -24489   205041          288
#> 29     TRUE    29            0    67            -19449   186020          214
#> 30    FALSE    30            0    67            -11219   174967          181
#>    transfers_out loaned_in loaned_out minutes goals_scored assists clean_sheets
#> 1              0         0          0       0            0       0            0
#> 2          77039         0          0      45            0       0            0
#> 3          36785         0          0      90            0       0            0
#> 4          18469         0          0      75            0       0            1
#> 5          17202         0          0      70            0       0            0
#> 6          15212         0          0      74            0       0            0
#> 7          17730         0          0       0            0       0            0
#> 8          27394         0          0       0            0       0            0
#> 9          24808         0          0       0            0       0            0
#> 10          7166         0          0       1            0       0            0
#> 11          2711         0          0      12            0       0            0
#> 12          2127         0          0      56            0       0            0
#> 13          2068         0          0      59            0       0            0
#> 14          2041         0          0       4            0       0            0
#> 15          2716         0          0      74            1       0            0
#> 16          1569         0          0      72            0       0            0
#> 17          6279         0          0      84            1       0            0
#> 18          2722         0          0      90            2       1            0
#> 19          6390         0          0      90            0       1            0
#> 20         34583         0          0      90            0       0            0
#> 21         89445         0          0       3            0       0            0
#> 22        116766         0          0      82            0       0            1
#> 23         56643         0          0      74            0       0            0
#> 24         36737         0          0      27            0       1            0
#> 25         30424         0          0      89            0       0            1
#> 26         27567         0          0       0            0       0            0
#> 27         60279         0          0      17            0       0            0
#> 28         24777         0          0       0            0       0            0
#> 29         19663         0          0       0            0       0            0
#> 30         11400         0          0       0            0       0            0
#>    goals_conceded own_goals penalties_saved penalties_missed yellow_cards
#> 1               0         0               0                0            0
#> 2               0         0               0                0            0
#> 3               3         0               0                0            0
#> 4               0         0               0                0            0
#> 5               1         0               0                0            0
#> 6               1         0               0                0            0
#> 7               0         0               0                0            0
#> 8               0         0               0                0            0
#> 9               0         0               0                0            0
#> 10              0         0               0                0            0
#> 11              0         0               0                0            0
#> 12              2         0               0                0            0
#> 13              0         0               0                0            0
#> 14              0         0               0                0            0
#> 15              2         0               0                0            1
#> 16              1         0               0                0            0
#> 17              3         0               0                0            0
#> 18              1         0               0                0            0
#> 19              1         0               0                0            0
#> 20              1         0               0                0            0
#> 21              0         0               0                0            0
#> 22              0         0               0                0            0
#> 23              1         0               0                0            0
#> 24              1         0               0                0            0
#> 25              0         0               0                0            1
#> 26              0         0               0                0            0
#> 27              0         0               0                0            0
#> 28              0         0               0                0            0
#> 29              0         0               0                0            0
#> 30              0         0               0                0            0
#>    red_cards saves bonus bps influence creativity threat ict_index ea_index
#> 1          0     0     0   0       0.0        0.0    0.0       0.0        0
#> 2          0     0     0   3       3.6       11.9    2.0       1.8        0
#> 3          0     0     0   4       6.2       13.7   51.0       7.1        0
#> 4          0     0     0  12       1.4       15.3   39.0       5.6        0
#> 5          0     0     0  11       8.4       19.7    0.0       2.8        0
#> 6          0     0     0  10       3.8        7.6   13.0       2.4        0
#> 7          0     0     0   0       0.0        0.0    0.0       0.0        0
#> 8          0     0     0   0       0.0        0.0    0.0       0.0        0
#> 9          0     0     0   0       0.0        0.0    0.0       0.0        0
#> 10         0     0     0   3       1.0        0.0    0.0       0.1        0
#> 11         0     0     0   3       6.2       11.3   17.0       3.5        0
#> 12         0     0     0   4       5.0       16.9    0.0       2.2        0
#> 13         0     0     0   9      11.2       24.3   37.0       7.3        0
#> 14         0     0     0   3       0.2        0.8    0.0       0.1        0
#> 15         0     0     0  20      37.8        4.4   19.0       6.1        0
#> 16         0     0     0   1       5.0        4.3    0.0       0.9        0
#> 17         0     0     2  26      35.6        3.3   20.0       5.9        0
#> 18         0     0     3  58      90.4       22.6   49.0      16.2        0
#> 19         0     0     0  17      26.6       24.8   11.0       6.2        0
#> 20         0     0     0  15       9.6       21.2    8.0       3.9        0
#> 21         0     0     0   6       2.2       10.1    0.0       1.2        0
#> 22         0     0     0   6       7.0        3.6    4.0       1.5        0
#> 23         0     0     0   2       4.4       11.8   14.0       3.0        0
#> 24         0     0     0  12       0.8        5.2    2.0       0.8        0
#> 25         0     0     0  10       5.6        5.3   35.0       4.6        0
#> 26         0     0     0   0       0.0        0.0    0.0       0.0        0
#> 27         0     0     0   1       0.0        0.3    2.0       0.0        0
#> 28         0     0     0   0       0.0        0.0    0.0       0.0        0
#> 29         0     0     0   0       0.0        0.0    0.0       0.0        0
#> 30         0     0     0   0       0.0        0.0    0.0       0.0        0
#>    open_play_crosses big_chances_created clearances_blocks_interceptions
#> 1                  0                   0                               0
#> 2                  0                   0                               0
#> 3                  0                   0                               1
#> 4                  0                   0                               0
#> 5                  1                   1                               2
#> 6                  0                   0                               0
#> 7                  0                   0                               0
#> 8                  0                   0                               0
#> 9                  0                   0                               0
#> 10                 0                   0                               0
#> 11                 0                   0                               1
#> 12                 1                   0                               0
#> 13                 0                   1                               0
#> 14                 0                   0                               0
#> 15                 0                   0                               2
#> 16                 0                   0                               1
#> 17                 0                   0                               2
#> 18                 0                   0                               0
#> 19                 0                   0                               1
#> 20                 0                   0                               0
#> 21                 0                   1                               0
#> 22                 0                   0                               1
#> 23                 0                   0                               0
#> 24                 0                   0                               0
#> 25                 0                   0                               0
#> 26                 0                   0                               0
#> 27                 0                   0                               0
#> 28                 0                   0                               0
#> 29                 0                   0                               0
#> 30                 0                   0                               0
#>    recoveries key_passes tackles winning_goals attempted_passes
#> 1           0          0       0             0                0
#> 2           1          1       0             0               20
#> 3           2          1       0             0               26
#> 4           3          1       0             0               43
#> 5           6          1       0             0               38
#> 6           0          0       1             0               41
#> 7           0          0       0             0                0
#> 8           0          0       0             0                0
#> 9           0          0       0             0                0
#> 10          0          0       0             0                0
#> 11          1          1       0             0                9
#> 12          3          1       0             0               18
#> 13          2          2       2             0               29
#> 14          0          0       0             0                3
#> 15          4          0       1             0               28
#> 16          3          0       0             0               28
#> 17          6          0       0             0               30
#> 18          3          1       0             0               61
#> 19          2          1       0             0               48
#> 20          1          1       2             0               56
#> 21          1          1       0             0                2
#> 22          4          0       0             0               33
#> 23          3          1       0             0               22
#> 24          0          0       0             0               24
#> 25          4          0       2             0               38
#> 26          0          0       0             0                0
#> 27          0          0       0             0                2
#> 28          0          0       0             0                0
#> 29          0          0       0             0                0
#> 30          0          0       0             0                0
#>    completed_passes penalties_conceded big_chances_missed
#> 1                 0                  0                  0
#> 2                18                  0                  0
#> 3                21                  0                  0
#> 4                39                  0                  0
#> 5                34                  0                  0
#> 6                36                  0                  0
#> 7                 0                  0                  0
#> 8                 0                  0                  0
#> 9                 0                  0                  0
#> 10                0                  0                  0
#> 11                6                  0                  0
#> 12               16                  0                  0
#> 13               24                  0                  0
#> 14                3                  0                  0
#> 15               22                  0                  0
#> 16               24                  0                  0
#> 17               22                  0                  0
#> 18               53                  0                  0
#> 19               39                  0                  0
#> 20               51                  0                  0
#> 21                2                  0                  0
#> 22               26                  0                  0
#> 23               16                  0                  1
#> 24               23                  0                  0
#> 25               33                  0                  0
#> 26                0                  0                  0
#> 27                1                  0                  0
#> 28                0                  0                  0
#> 29                0                  0                  0
#> 30                0                  0                  0
#>    errors_leading_to_goal errors_leading_to_goal_attempt tackled offside
#> 1                       0                              0       0       0
#> 2                       0                              0       2       0
#> 3                       0                              0       2       1
#> 4                       0                              0       2       0
#> 5                       0                              0       5       0
#> 6                       0                              0       2       0
#> 7                       0                              0       0       0
#> 8                       0                              0       0       0
#> 9                       0                              0       0       0
#> 10                      0                              0       0       0
#> 11                      0                              0       0       1
#> 12                      0                              0       0       0
#> 13                      0                              1       1       0
#> 14                      0                              0       0       0
#> 15                      0                              0       2       0
#> 16                      0                              0       3       1
#> 17                      0                              0       2       1
#> 18                      0                              0       1       0
#> 19                      0                              0       2       1
#> 20                      0                              0       2       1
#> 21                      0                              0       1       0
#> 22                      0                              0       1       0
#> 23                      0                              0       0       0
#> 24                      0                              0       0       0
#> 25                      0                              0       1       0
#> 26                      0                              0       0       0
#> 27                      0                              0       1       0
#> 28                      0                              0       0       0
#> 29                      0                              0       0       0
#> 30                      0                              0       0       0
#>    target_missed fouls dribbles element fixture opponent_team    playername
#> 1              0     0        0     300       6            11 Jesse Lingard
#> 2              0     0        1     300      11             3 Jesse Lingard
#> 3              1     0        1     300      26            17 Jesse Lingard
#> 4              2     0        2     300      32             4 Jesse Lingard
#> 5              0     2        0     300      49            18 Jesse Lingard
#> 6              1     1        2     300      59            20 Jesse Lingard
#> 7              0     0        0     300      69            19 Jesse Lingard
#> 8              0     0        0     300      77            15 Jesse Lingard
#> 9              0     0        0     300      84             6 Jesse Lingard
#> 10             0     0        0     300      97             8 Jesse Lingard
#> 11             0     0        0     300     102             2 Jesse Lingard
#> 12             0     3        1     300     118            13 Jesse Lingard
#> 13             0     1        0     300     126             7 Jesse Lingard
#> 14             0     0        0     300     140            16 Jesse Lingard
#> 15             0     3        0     300     145             1 Jesse Lingard
#> 16             0     2        0     300     158             9 Jesse Lingard
#> 17             1     1        2     300     165            12 Jesse Lingard
#> 18             0     1        3     300     173             5 Jesse Lingard
#> 19             0     0        0     300     187            10 Jesse Lingard
#> 20             0     1        2     300     197             2 Jesse Lingard
#> 21             0     0        0     300     208            15 Jesse Lingard
#> 22             1     1        0     300     219            17 Jesse Lingard
#> 23             1     2        0     300     226             3 Jesse Lingard
#> 24             0     0        0     300     235             4 Jesse Lingard
#> 25             0     2        1     300     247            11 Jesse Lingard
#> 26             0     0        0     300     253             9 Jesse Lingard
#> 27             0     1        0     300     268            12 Jesse Lingard
#> 28             0     0        0     300     273             7 Jesse Lingard
#> 29             0     0        0     300     286            16 Jesse Lingard
#> 30             0     0        0     300     291             1 Jesse Lingard