get_player_hist

This function fetches historical summary information on selected players given player ID(s).

get_player_hist(playerid = NULL)

Arguments

playerid
The player's ID. Can be found using get_player_id(). You can list multiple , e.g. with c(). If blank, the function fetches all current players - note that takes a while to load.

Examples

get_player_hist(playerid=300)

#>     id season_name element_code start_cost end_cost total_points minutes
#> 1 3830     2011/12       109322         55       55            0       0
#> 2 4367     2012/13       109322         45       43            0       0
#> 3 5184     2013/14       109322         45       42            0       0
#> 4 5800     2014/15       109322         45       43            1      23
#> 5 6638     2015/16       109322         45       43           76    1632
#> 6 7223     2016/17       109322         60       56           60    1360
#> 7 7900     2017/18       109322         60       60          128    1807
#>   goals_scored assists clean_sheets goals_conceded own_goals penalties_saved
#> 1            0       0            0              0         0               0
#> 2            0       0            0              0         0               0
#> 3            0       0            0              0         0               0
#> 4            0       0            0              0         0               0
#> 5            4       1            9             13         0               0
#> 6            1       3            7             12         0               0
#> 7            8       6            9             18         0               0
#>   penalties_missed yellow_cards red_cards saves bonus bps influence creativity
#> 1                0            0         0     0     0   0       0.0        0.0
#> 2                0            0         0     0     0   0       0.0        0.0
#> 3                0            0         0     0     0   0       0.0        0.0
#> 4                0            0         0     0     0   3       0.0        0.0
#> 5                0            4         0     0     5 268       0.0        0.0
#> 6                0            3         0     0     3 227     229.4      522.3
#> 7                0            2         0     0    10 450     526.6      519.1
#>   threat ict_index ea_index season    playername
#> 1    0.0       0.0        0      6 Jesse Lingard
#> 2    0.0       0.0        0      7 Jesse Lingard
#> 3    0.0       0.0        0      8 Jesse Lingard
#> 4    0.0       0.0        2      9 Jesse Lingard
#> 5    0.0       0.0      318     10 Jesse Lingard
#> 6  439.0     118.0        0     11 Jesse Lingard
#> 7  711.0     175.3        0     12 Jesse Lingard

get_player_hist(playerid=get_player_id("Eden Hazard"))

#> Warning: cannot open URL 'https://fantasy.premierleague.com/drf/element-summary/Eden Hazard': HTTP status was '404 Not Found'

#> Error in open.connection(con, "rb"): kan ikke åbne forbindelsen