get_player_info

This function fetches selected summary information on selected players given full player name(s).

get_player_info(name = NULL, season = NULL)

Arguments

name
The player's full name, as listed on the official Fantasy Premier League site (for instance: "Richarlison de Andrade", not "Richarlison"). If blank, the function fetches all players.

season
To retrieve prior season data, enter a 2-digit year corresponding to the start of the historical FPL season of interest (e.g. '17' for the 2017/2018 season).

Examples

get_player_info(name=c("Jesse Lingard","Virgil van Dijk"))

#>   chance_of_playing_next_round chance_of_playing_this_round   code
#> 1                          100                          100  97032
#> 2                          100                          100 109322
#>   cost_change_event cost_change_event_fall cost_change_start
#> 1                 0                      0                 0
#> 2                 0                      0                -2
#>   cost_change_start_fall dreamteam_count element_type ep_next ep_this
#> 1                      0               2            2     3.2     4.1
#> 2                      2               0            3     0.7     1.7
#>   event_points first_name form  id in_dreamteam news
#> 1            9     Virgil  3.1 183         TRUE     
#> 2            6      Jesse  1.2 241        FALSE     
#>                    news_added now_cost      photo points_per_game second_name
#> 1 2019-10-27T20:00:16.725016Z       65  97032.jpg             4.7    van Dijk
#> 2 2020-03-13T10:30:19.085305Z       63 109322.jpg             1.5     Lingard
#>   selected_by_percent special squad_number status team team_code total_points
#> 1                42.4   FALSE           NA      a   10        14          178
#> 2                 0.7   FALSE           NA      a   12         1           32
#>   transfers_in transfers_in_event transfers_out transfers_out_event value_form
#> 1      2145436                  0       2817026                   0        0.5
#> 2        64075                  0        129253                   0        0.2
#>   value_season web_name minutes goals_scored assists clean_sheets
#> 1         27.4 van Dijk    3420            5       2           15
#> 2          5.1  Lingard     923            1       0            1
#>   goals_conceded own_goals penalties_saved penalties_missed yellow_cards
#> 1             33         0               0                0            1
#> 2             14         0               0                0            3
#>   red_cards saves bonus bps influence creativity threat ict_index
#> 1         0     0    14 804     930.2      152.4  467.0     154.8
#> 2         0     0     0  99     109.0      278.2  297.0      68.2
#>   influence_rank influence_rank_type creativity_rank creativity_rank_type
#> 1             12                   2             236                   62
#> 2            375                 157             151                   95
#>   threat_rank threat_rank_type ict_index_rank ict_index_rank_type
#> 1          80                4             61                   6
#> 2         135               73            241                 117
#>        playername
#> 1 Virgil van Dijk
#> 2   Jesse Lingard